+++ src/components/Landing/Search.jsx
@@ -1,5 +1,0 @@
-function Search() {
-  return <div>{/* 검색창 */}</div>;
-}
-
-export default Search;