+++ src/styles/element.styled.js
@@ -1,5 +1,5 @@
 import { styled } from '@mui/styles';
-import { Container } from '@mui/material';
+import { Container,Paper,Grid } from '@mui/material';
 import { Link } from 'react-router-dom';
 import { theme } from './theme';
 
@@ -16,4 +16,14 @@ export const StyledContainer = styled(Container)({
 export const StyledLink = styled(Link)({
   color: (props) => (props.color ? props.color : theme.palette.primary.main),
   textDecoration: 'none',
+});
+
+export const StyledPaper = styled(Paper)({
+  padding: '3rem',
+  alignItems: 'center',
+  justifyContent: 'center',
+});
+
+export const StyledGrid = styled(Grid)({
+  height: '60vh',
 });